Switch back to mmap s16; try out some BLAS which does not work

+++ arm/capture.h
@@ -5,7 +5,7 @@ struct CaptureContextTag;
 typedef struct CaptureContextTag CaptureContext;
 
 
-typedef float sample_t;
+typedef int16_t sample_t;
 
 
 CaptureContext *capture_init(void);
